+++ xarray/core/variable.py
@@ -660,7 +660,7 @@ class Variable(AbstractArray, NdimSizeLenMixin, VariableArithmetic):
         return dims, BasicIndexer(key), None
 
     def _validate_indexers(self, key):
-        """ Make sanity checks """
+        """Make sanity checks"""
         for dim, k in zip(self.dims, key):
             if isinstance(k, BASIC_INDEXING_TYPES):
                 pass
@@ -715,7 +715,7 @@ class Variable(AbstractArray, NdimSizeLenMixin, VariableArithmetic):
         return dims, OuterIndexer(tuple(new_key)), None
 
     def _nonzero(self):
-        """ Equivalent numpy's nonzero but returns a tuple of Varibles. """
+        """Equivalent numpy's nonzero but returns a tuple of Varibles."""
         # TODO we should replace dask's native nonzero
         # after https://github.com/dask/dask/issues/1076 is implemented.
         nonzeros = np.nonzero(self.data)
@@ -2140,16 +2140,17 @@ class Variable(AbstractArray, NdimSizeLenMixin, VariableArithmetic):
         Apply reduction function.
         """
         windows = {k: v for k, v in windows.items() if k in self.dims}
-        if not windows:
-            return self.copy()
 
         if keep_attrs is None:
-            keep_attrs = _get_keep_attrs(default=False)
+            keep_attrs = _get_keep_attrs(default=True)
 
         if keep_attrs:
             _attrs = self.attrs
         else:
             _attrs = None
+
+        if not windows:
+            return self._replace(attrs=_attrs)
 
         reshaped, axes = self._coarsen_reshape(windows, boundary, side)
         if isinstance(func, str):